requirements.txt
alembic==1.0.5
Babel==2.6.0
backcall==0.1.0
chardet==3.0.4
Click==7.0
dbfread==2.0.7
decorator==4.3.0
dominate==2.3.5
elasticsearch==6.3.1
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
leather==0.3.3
Mako==1.0.7
MarkupSafe==1.1.0
openpyxl==2.5.12
packaging==19.0
parsedatetime==2.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pipdeptree==0.13.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-slugify==1.2.6
pytimeparse==1.1.8
pytz==2018.7
six==1.12.0
SQLAlchemy==1.2.15
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.1
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
